+++ src/routes/api/categories.json.ts
@@ -1,0 +1,12 @@
+export const get = async () => {
+  return {
+    body: {
+        pd: {
+            title: 'Personal Development',
+        },
+        dev: {
+            title: 'Web/App Development',
+        }
+    }
+  }
+}