+++ src/routes/blog/index.svelte
@@ -1,18 +1,21 @@
 <script context="module">
 export const load = async ({ fetch }) => {
-  const posts = await fetch('/api/posts.json')
-  const allPosts = await posts.json()
+  const getPosts = await fetch('/api/posts.json')
+  const getCategories = await fetch('/api/categories.json')
+  const categories = await getCategories.json()
+  const posts = await getPosts.json()
 
   return {
     props: {
-      posts: allPosts
+      posts,
+      categories,
     }
   }
 }
 </script>
 
 <script>
-    export let posts;
+    export let posts, categories;
 </script>
 
 <h1>Blog</h1>
@@ -24,7 +27,7 @@ export const load = async ({ fetch }) => {
           {post.meta.title}
         </a>
       </h2>
-      Published {post.meta.date}
+      {categories[post.meta.category].title}, {new Date(post.meta.date).toLocaleDateString('de-DE')}
     </li>
   {/each}
 </ul>